PoolStore moved to store/sqlite_store.py
